Fix paths for packaging.

+++ audio.py
@@ -10,18 +10,22 @@ import re
 from PyQt4 import QtGui
 from PyQt4.uic import compileUiDir
 
-compileUiDir(os.path.join(os.path.dirname(__file__), 'audio', 'resources'))
-for file_ in glob.glob(os.path.join(os.path.dirname(__file__), 'audio', 'resources', '*.py')):
-    infile = open(file_)
-    outfile = open(os.path.join(os.path.dirname(__file__), 'audio', 'ui', os.path.basename(file_)), 'w')
-    for i in infile:
-        n = re.sub(r'(#\s*Created:.*)', '#', i)
-        a = re.sub(r'(#\s*by:.*)', '#', n)
-        outfile.write(a)
-    infile.close()
-    outfile.close()
-    if os.path.exists(file_):
-        os.remove(file_)
+if getattr(sys, 'frozen', False):
+    basedir = sys._MEIPASS
+else:
+    basedir = os.path.dirname(__file__)
+    compileUiDir(os.path.join(basedir, 'audio', 'resources'))
+    for file_ in glob.glob(os.path.join(basedir, 'audio', 'resources', '*.py')):
+        infile = open(file_)
+        outfile = open(os.path.join(basedir, 'audio', 'ui', os.path.basename(file_)), 'w')
+        for i in infile:
+            n = re.sub(r'(#\s*Created:.*)', '#', i)
+            a = re.sub(r'(#\s*by:.*)', '#', n)
+            outfile.write(a)
+        infile.close()
+        outfile.close()
+        if os.path.exists(file_):
+            os.remove(file_)
 
 from audio.ui.mainwindowform import MainWindow
